feat: use TS variant of `class-methods-use-this`
By default, this does not deviate in behavior from the original rule.
It does, however, provide users with additional configuration options
to disable errors when `this` is not used in classes that
implement interfaces or methods that explicitly override parent methods.

The configuration we provide has been left at the default, though,
since those rules are arguably over-permissive in the pursuit of not
requiring type information.

Closes #344.